fix: pin asdf toolchain versions so all tests pass
Match the .tool-versions Elixir version to what is already in the
mix.exs file (Elixir 1.10).

Add corresponding Erlang OTP version.

Add a reasonable Java runtime for the PlantUML tasks to pass muster.
This will be especially helpful to those of us who have not kept in
touch with recent developments in the Java world.