test: instrument tests to show diffs 1.10 to 1.13
The arch map is a large data structure when run on the clean_mixer
project itself. I'm trying to figure out what differences show up in
that data structure when run under Elixir 1.13 versus 1.10.

This commit is merely a step along the way and does not have it figured
out yet.