fix: get all tests passing by adding explicit refs
Turns out the compiler optimizations done since Elixir 1.10.4 made it so
that when module A references module B by simply declaring an alias to
B, the compiler doesn't count it as a first-class reference unless code
in A actually calls code in B at some point.

This commit includes the extensive scaffolding used to narrow down the
problem. My intention is to tag this commit so I can get back to the
scaffolding should I need it in the future.

The next commits will show the work of removing the scaffolding and
getting back to the minimum changes necessary to get all tests passing
again in newer versions of Elixir.